LayoutTests/media/modern-media-controls/media-controller/media-controller-click-on-video-background-should-pause-expected.txt

Testing that clicking on a video's background pauses it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS media.paused is false

Clicking on video background
PASS media.paused is true

PASS successfullyParsed is true

TEST COMPLETE
